added conftest for shared test resources
