+++ SmartApp.CV/online/tests/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+from online.SDK.face_client import Facepp_Client
+
+@pytest.fixture
+def client():
+    client = Facepp_Client()
+    client.setParamsDetect()
+    return client
